Original test file
This is the original test file that is manually changed by a user before it is copied for testing the whole program. 
The format must remain the same
Number of lines does not affect the program, as long as each line is written in dictionary format.
Also, each pin numbers and account numbers must be different from each other.